+++ legends/gameobjects.py
@@ -5,11 +5,15 @@
 from types import MethodType
 from legends.utils.objrelations import Managed, OneToOne
 #pylint: disable-next=no-name-in-module
-from legends.constants import GSSkill, GSCharacter, GSLevel, GSGear, GSRank
-from legends.constants import DESCRIPTIONS, RARITIES
+from legends.constants import (
+    GSSkill, GSCharacter, GSLevel, GSGear, GSRank, GSBaseStat, GSGearLevel
+)
+from legends.constants import DESCRIPTIONS, RARITIES, PART_STAT_VALUES
+from legends.stats import Stats
 
 __all__ = [
-    'levelFromXP', 'tokensNeeded', 'Gear', 'Particle', 'GearSlot', 'PartSlot',
+    'levelFromXP', 'xpFromLevel', 'tokensNeeded', 'getCharStats',
+    'getGearStats', 'getPartStats', 'Gear', 'Particle', 'GearSlot', 'PartSlot',
     'Character', 'Skill', 'Roster'
 ]
 
@@ -34,6 +38,19 @@ def levelFromXP(xp, rarity='Common'):
         raise ValueError(repr(xp) + ' could not be converted from XP to level')
     return level
 
+def xpFromLevel(level, rarity='Common'):
+    """Calculates the minimum XP of the character from its level.
+
+    Args:
+        level (int): The level of the character.
+        rarity (str): The rarity of the character.
+
+    Returns:
+        int: The minimum possible XP the character could have.
+
+    """
+    return GSLevel[rarity + '_' + str(level)]['Experience']
+
 def tokensNeeded(rarity, rank):
     """Returns the number of tokens needed by a character of the given
     rarity and rank to move up to the next rank.
@@ -42,6 +59,89 @@ def tokensNeeded(rarity, rank):
         return 0
     # pylint: disable-next=undefined-variable
     return GSRank['{}_{}'.format(rarity, rank)]['RequiredTokenCount']
+
+def getCharStats(nameID, rank, level):
+    """Calculates a character's naked stats from its nameID, rank, and
+    level.
+
+    Args:
+        nameID (str): The name ID of the character, as it appears in
+            `GSCharacter`.
+        rank (int): The rank of the character.
+        level (int): The level of the character.
+
+    Returns:
+        dict: A dictionary mapping stat names, as they appear in
+            `GSBaseStat`, to stat values.
+
+    """
+    rarity = GSCharacter[nameID]['Rarity']
+    stats = {}
+    for statName, data in GSBaseStat.items():
+        m = data['MinValue'] #pylint: disable=invalid-name
+        M = data['MaxValue'] #pylint: disable=invalid-name
+        t = GSCharacter[nameID][statName] #pylint: disable=invalid-name
+        baseStat = m + t * (M - m)
+        try:
+            levelMod = GSLevel[rarity + '_' + str(level)][
+                    statName + 'Modifier'
+                ]
+            rankMod = GSRank[rarity + '_' + str(rank)][
+                    statName + 'Modifier'
+                ]
+        except KeyError:
+            levelMod = 1
+            rankMod = 1
+        statVal = baseStat * levelMod * rankMod
+        stats[statName] = statVal
+    return stats
+
+def getGearStats(gearID, level):
+    """Calculates a gear's stats from its gear ID and level.
+
+    Args:
+        gearID (str): The gear ID as it appears in `GSGear`.
+        level (int): The level of the gear. Level is 1-based and
+            includes rarity. For example, a Level 17 gear piece displays
+            in game as "Epic, Level 2".
+
+    Returns:
+        dict: A dictionary mapping stat names, as they appear in
+            `GSBaseStat`, to stat values.
+
+    """
+    stats = {statName: 0 for statName in GSBaseStat}
+    gearLevelID = '[{}, {}]'.format(gearID, level)
+    numStats = GSGearLevel[gearLevelID]['m_StatBrancheCount']
+    for i in range(numStats):
+        data = GSGear[gearID]['m_Stats'][i]
+        statName = data['m_Type']
+        statBase = data['m_BaseValue']
+        statIncr = data['m_IncreaseValue']
+        statVal = statBase + (level - 10 * i) * statIncr
+        stats[statName] += statVal
+    return stats
+
+def getPartStats(rarity, level, statList):
+    """Calculates a particle's stats from its rarity, level, and list of
+    stat names.
+
+    Args:
+        rarity (str): The particle's rarity.
+        level (int): The particle's level.
+        statList (list of str): The stat names on the particle.
+
+    Returns:
+        dict: A dictionary mapping stat names, as they appear in
+            `GSBaseStat`, to stat values.
+
+    """
+    stats = {statName: 0 for statName in GSBaseStat}
+    for statName in statList:
+        stats[statName] = (
+            PART_STAT_VALUES[statName][rarity][level - 1]
+        )
+    return stats
 
 class Gear(Managed):
     """A piece of gear in STL.
@@ -52,17 +152,41 @@ class Gear(Managed):
         level (int): The level of the gear. Level is 1-based and
             includes rarity. For example, a Level 17 gear piece displays
             in game as "Epic, Level 2".
+        stats (Stats): The Stats object that stores the gear's total
+            stats.
 
     """
 
     def __init__(self, gearID, level):
         self.gearID = gearID
-        self.level = level
+        self._level = level
+        self.stats = Stats()
+        self.updateStats()
+
+    @property
+    def level(self):
+        """int: The level of the gear. Level is 1-based and includes
+        rarity. For example, a Level 17 gear piece displays in game as
+        "Epic, Level 2". Modifying this property also calls the
+        `updateStats` method.
+        """
+        return self._level
+
+    @level.setter
+    def level(self, value):
+        self._level = value
+        self.updateStats()
 
     @property
     def slot(self):
         """The slot index into which the gear must be placed."""
         return GSGear[self.gearID]['m_Slot']
+
+    def updateStats(self):
+        """Updates the `stats` attribute.
+
+        """
+        self.stats.update(getGearStats(self.gearID, self.level))
 
     def __repr__(self):
         return '<' + repr(self.gearID) + ', level ' + repr(self.level) + '>'
@@ -76,17 +200,57 @@ class Particle(Managed):
         rarity (str): The particle's rarity.
         level (str): The particle's level.
         locked (bool): True if the particle is locked.
-        statName (list of str): The names, as they appear in GSBaseStat,
-            of the stats that are on the particle.
 
     """
 
     def __init__(self, name, rarity, level, locked=False):
         self.name = name
         self.rarity = rarity
-        self.level = level
+        self._level = level
         self.locked = locked
-        self.statNames = []
+        self._statNames = [None] * 4
+        self.stats = Stats()
+
+    @property
+    def level(self):
+        """str: The particle's level. Modifying this property also calls
+        the `updateStats` method.
+        """
+        return self._level
+
+    @level.setter
+    def level(self, value):
+        self._level = value
+        self.updateStats()
+
+    @property
+    def statNames(self):
+        """tuple of str: The names, as they appear in GSBaseStat, of the
+        stats that are on the particle.
+        """
+        return tuple(self._statNames)
+
+    def setStatName(self, index, statName):
+        """Changes the `statNames` property by setting the value at the
+        given index to the given stat name.
+
+        Args:
+            index (int): The 0-based index of the value in the
+                `statNames` property to change.
+            statName (str): The stat name to assign to the given index.
+
+        """
+        self._statNames[index] = statName
+        self.updateStats()
+
+    def updateStats(self):
+        """Updates the `stats` attribute.
+
+        """
+        statList = [
+            statName for statName in self.statNames if statName is not None
+        ]
+        self.stats.update(getPartStats(self.rarity, self.level, statList))
 
     def __repr__(self):
         return (
@@ -138,18 +302,21 @@ class Character():
     Attributes:
         nameID (str): The in-data name of the character. Should match a
             key in `GSCharacter`.
-        rank (int): The rank of the character.
-        tokens (int): The number of tokens obtained toward ranking the
-            character.
-        xp (int): The xp the character has.
+        stats (Stats): The Stats object that stores the character's
+            naked stats (i.e. the stats they would have without any gear
+            or particles).
         skills (dict): A dictionary mapping skill IDs (found in
             `GSSkill`) to Skill objects.
 
     """
     def __init__(self, nameID, rank=1, xp=0):
-        self.nameID = nameID
-        self.rank = rank
-        self.xp = xp
+        self._data = {
+            'nameID': nameID,
+            'rank': rank,
+            'xp': xp
+        }
+        self.stats = Stats()
+        self.updateStats()
         self.skills = {
             skillID: Skill(skillID)
             for skillID in GSCharacter[self.nameID]['SkillIDs']
@@ -162,14 +329,53 @@ class Character():
             self.partSlots.append(PartSlot(self, slot))
 
     @property
+    def nameID(self):
+        """str: The in-data name of the character. Should match a key in
+        `GSCharacter`.
+        """
+        return self._data['nameID']
+
+    @property
     def name(self):
         """The in-game display name of the character."""
         return DESCRIPTIONS[GSCharacter[self.nameID]['Name']]
 
     @property
+    def rank(self):
+        """int: The rank of the character. Modifying this property also
+        calls the `updateStats` method.
+        """
+        return self._data['rank']
+
+    @rank.setter
+    def rank(self, value):
+        self._data['rank'] = value
+        self.updateStats()
+
+    @property
+    def xp(self):
+        """int: The xp of the character. Modifying this property also
+        calls the `updateStats` method.
+        """
+        return self._data['xp']
+
+    @xp.setter
+    def xp(self, value):
+        self._data['xp'] = value
+        self.updateStats()
+
+    @property
     def level(self):
-        """int: The level of the character."""
+        """int: The level of the character. Setting this to a value
+        different from its current value will change the `xp` property
+        to minimum xp required to attain the new level.
+        """
         return levelFromXP(self.xp)
+
+    @level.setter
+    def level(self, value):
+        if self.level != value:
+            self.xp = xpFromLevel(value, self.rarity)
 
     @property
     def rarity(self):
@@ -192,6 +398,12 @@ class Character():
         reach the next rank.
         """
         return tokensNeeded(self.rarity, self.rank)
+
+    def updateStats(self):
+        """Updates the `stats` attribute.
+
+        """
+        self.stats.update(getCharStats(self.nameID, self.rank, self.level))
 
     def __repr__(self):
         return (
@@ -270,3 +482,36 @@ class Roster():
     def containsPart(self):
         """OneToOne: The inverse of `inPartSlot`."""
         return self.inPartSlot.inverse
+
+    def charStats(self, nameID):
+        """Constructs and returns a Stats object containing the total
+        stats (including gear and particles) of the character with the
+        given name ID.
+
+        Args:
+            nameID (str): The name ID of the character whose stats to
+                build.
+
+        Returns:
+            Stats: The constructed Stats object.
+
+        """
+        char = self.chars[nameID]
+        nakedStats = char.stats
+        gears = (
+            self.containsGear[gearSlot] for gearSlot in char.gearSlots
+            if gearSlot in self.containsGear
+        )
+        gearStats = sum(
+            (gear.stats for gear in gears if gear is not None),
+            Stats()
+        )
+        parts = (
+            self.containsPart[partSlot] for partSlot in char.partSlots
+            if partSlot in self.containsPart
+        )
+        partStats = sum(
+            (part.stats for part in parts if part is not None),
+            Stats()
+        )
+        return nakedStats + gearStats + partStats